Composer, Spelling, Docstrings, PSR-4
