+++ .gitignore
@@ -1,0 +1,7 @@
+vendor
+bin
+composer.lock
+coverage
+coverage.xml
+tests/files
+.idea
