+++ composer.json
@@ -1,0 +1,26 @@
+{
+    "name": "platformcommunity/bunnycdn-storage",
+    "description": "Library for BunnyCDN Storage",
+    "license": "N/A (Pending)",
+    "authors": [
+        {
+            "name": "Dejan Grofelnik Pelzel",
+            "email": "hello@bunnycdn.com"
+        }
+    ],
+    "require": {
+        "php": ">=5.6.0"
+    },
+    "require-dev": {
+    },
+    "autoload": {
+        "psr-4": {
+            "BunnyWay\\BunnyCDN\\": "src/"
+        }
+    },
+    "extra": {
+        "branch-alias": {
+            "dev-master": "1.0-dev"
+        }
+    }
+}
